make the verifymembers model look nicer, fix the indentation. please, have a look at the ProgrammingGuideline:  * always use 4 spaces per indentation level, no tabs.  * mysql keywords go all uppercase, and ideally multi-line.  * method names in lowerCamelCase  * local variable names in $lowercase_with_underscores. This makes it easier for others to read. Sorry for complaining.
The main thing was to use the built-in sql methods "singleLookup()" and "bulkLookup()" defined in RoxModelBase. (maybe "loadRow()" and "loadRows()" would be better names for them..)

I was not able to fix the problem with the missing POST values - I found out it is the PT PPostHandler's fault, which does a redirect. Obviously, this will flush the POST values down the toilet. I am trying to analyse.

git-svn-id: file:///home/jaliste/hcvol/hcvol/trunk@5040 4f49b16c-3c57-40ef-8e22-7452baa6c33a