notify: fixing broken function updateNote of the note entity
as the function was, it would only have allowed to set most parameters
in one way. f.ex. updateNote(false) wouldn't have unchecked the note.

furthermore there was a typo in translationparams so any update of
those would effectively have emptied the field.

fortunately this never led to a bug because the only call of the
function was luckily updateNote(true) :P