fix the "report bug" link in the footer to point to the feedback page. see http://trac.bewelcome.org/ticket/1546 .
the feedback page accepts an extra GET-parameter RequestURI which is
added to the feedback once it is sent. this gives information about on
which page the user clicked on the "report bug" link.

further the version information from revision.txt is also appended to
the feedback text.

to retrieve this version information i added a new function
getVersionInfo to RoxComponentBase. I am not sure this is the right
place for such a function. Another possible option would be
RoxModelBase. Review needed.