deprecate MOD_user::update{Database,Session}OnlineCounter in favor of MOD_online
MOD_online is a rewritten clean module which keeps the online counters
of guests and members up-to-date. The functionality and effective code
is largely the same, with AFAICS only two exceptions:

- lastactivity column is NOT updated anymore (consider it
  privacy-enhancement)
- $_SESSION['Params'] is NOT fetched anew. But instantiating a
  RoxModelBase calls MOD_params::get()->loadParams(), which outdates
  the need for it

Further the code checks now in the frontrouter, whether a member has
been banned, and then logs him out to kill the session.