Fixing search by age
Members who give a fake birth date making them more than 120 years old are prensently not displayed in search results... There are more of them than one would think, just choosing a funny birth date without thinking it will be an issue. I changed the condition so that setting the min age to 18 and max age to 120 (the widest range available in search parameters, and default one) is the same as setting no condition.
Not sure when the $vars['search-age-minimum'] and $vars['search-age-maximum'] can be set and equal to 0, so I left that condition.